// src/server/api/models/crm/models/KycDocumentRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing a kyc document. The fields marked with an * means that they are required.
 */
export type KycDocumentRequest = {
    /**
     * A unique identifier for the client that is generated by the crm
     */
    clientId: string;
    /**
     * A unique identifier for the contact that is generated by the crm
     */
    contactId?: string;
    /**
     * The record category submitted.
     */
    recordType: KycDocumentRequest.recordType;
    /**
     * The type of ID document submitted.
     */
    idType?: KycDocumentRequest.idType;
    /**
     * An optional ID no. This is required for electronic ID validation workflows
     */
    idNo?: string;
    /**
     * Optionally specify the authority that issued the document
     */
    idIssuer?: string;
    /**
     * The expiration date of the ID document
     */
    idExpDate?: string;
    /**
     * The type of file
     */
    fileType?: string;
    /**
     * The name of the file
     */
    fileName?: string;
    /**
     * The size of the uploaded file
     */
    fileSize?: number;
    /**
     * Any notes on the document
     */
    notes?: string;
    /**
     * A base 64 encoded representation of the fileContent. To delete this value when patching these records set the value to 'null'
     */
    fileContent?: string;
};

export namespace KycDocumentRequest {

    /**
     * The record category submitted.
     */
    export enum recordType {
        PICTURE = 'PICTURE',
        SIGNATURE = 'SIGNATURE',
        IDENTITY = 'IDENTITY',
        ADDRESS = 'ADDRESS',
        OTHERS = 'OTHERS',
    }

    /**
     * The type of ID document submitted.
     */
    export enum idType {
        PASSPORT = 'PASSPORT',
        NATIONAL_ID = 'NATIONAL_ID',
        DRIVERS_LICENSE = 'DRIVERS_LICENSE',
        COMPANY_ID = 'COMPANY_ID',
        VOTER_ID = 'VOTER_ID',
        BIRTH_CERTIFICATE = 'BIRTH_CERTIFICATE',
        RESIDENCE_PERMIT = 'RESIDENCE_PERMIT',
        WORK_PERMIT = 'WORK_PERMIT',
        STUDENT_ID = 'STUDENT_ID',
        SOCIAL_SECURITY_CARD = 'SOCIAL_SECURITY_CARD',
        CERTIFICATE_OF_CITIZENSHIP = 'CERTIFICATE_OF_CITIZENSHIP',
        CERTIFICATE_OF_INCORPORATION = 'CERTIFICATE_OF_INCORPORATION',
        CERTIFICATE_OF_LIQUIDATION = 'CERTIFICATE_OF_LIQUIDATION',
        CERTIFICATE_OF_NATURALIZATION = 'CERTIFICATE_OF_NATURALIZATION',
        CERTIFICATE_OF_RESIDENCE = 'CERTIFICATE_OF_RESIDENCE',
        UTILITY_BILL = 'UTILITY_BILL',
        BANK_STATEMENT = 'BANK_STATEMENT',
        PAYSLIP = 'PAYSLIP',
        TAX_RETURN = 'TAX_RETURN',
        OTHER_DOCUMENTS = 'OTHER_DOCUMENTS',
    }


}